+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- The `capsicum::casper::Service` trait's `SERVICE_NAME` field must now be
+  defined as a `&'static CStr`.  The [cstr](https://crates.io/crates/cstr)
+  crate can help.
+  ([#49](https://github.com/dlrobertson/capsicum-rs/pull/49))
+
 ### Fixed
 
 - Fixed cross-building the documentation.